refactor: restructure project with core and extensions
