+++ .env
@@ -1,15 +1,10 @@
-TUYA_API_KEY    = "n95maffcb943oym8m7mr"
-TUYA_API_SECRET = "c26e537f971d4adf960ea0c47c48f693"
-TUYA_API_REGION = "eu"
-
-MQTT_BROKER_HOST = "127.0.0.1"
-MQTT_BROKER_PORT = "1883"
-# MQTT_USERNAME    = "timas"
-# MQTT_PASSWORD    = "dubna123"
-
-# Service config files
-TUYA2MQTT_DEV_CONF_FILE = "/home/tsmolyanin/wk/ttmp/tinytuya2mqtt/devices.json"
-TUYA2MQTT_LOCAL_SCAN_FILE = "/home/tsmolyanin/wk/ttmp/tinytuya2mqtt/local_scan.json"
-
-# Polling interval (seconds)
-TUYA2MQTT_POLL_INTERVAL = "2"
+TUYA_API_KEY="n95maffcb943oym8m7mr"
+TUYA_API_SECRET="c26e537f971d4adf960ea0c47c48f693"
+TUYA_API_REGION="eu"
+MQTT_BROKER_HOST="127.0.0.1"
+MQTT_BROKER_PORT="1883"
+# MQTT_USERNAME="timas"
+# MQTT_PASSWORD="dubna123"
+TUYA2MQTT_DEV_CONF_FILE="/home/tsmolyanin/wk/ttmp/tinytuya2mqtt/devices.json"
+TUYA2MQTT_LOCAL_SCAN_FILE="/home/tsmolyanin/wk/ttmp/tinytuya2mqtt/local_scan.json"
+TUYA2MQTT_POLL_INTERVAL="2"
