--- adapters/__init__.py
@@ -1,3 +1,0 @@
-from .tiny_transport import TinyLocalTransport
-from .tiny_cloud import TinyCloud, Cloud
-from .tuya_constants import *
