+++ core/domain_event_types.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+@dataclass
+class StatusPolled:
+    dev_id: str
+    dps: dict
+
+@dataclass
+class MqttCommand:
+    topic: str
+    payload: str
