+++ core/utility_functions.py
