+++ extensions/homie/common/templates/rgb_plug.json
